Update README.md
Added sed command to remove the continue 2 in `hyprdots/Scripts/restore_cfg.sh`. This line was preventing the restoration because it is inside an if statement that checks whether a particular package is installed. The failure of the check resulted in the execute of `continue 2` which skipped the restoration of configs.